added database to explore
